moved resources into "private" folders
