+++ .gitignore
@@ -22,4 +22,4 @@ _configs.php
 .idea
 *.sql
 tmp
-media/_notes
+_media/_notes
