+++ admin/listReservations.php
@@ -1,9 +1,9 @@
 <?php
     require "../login/denyNoAdmin.php";
     $page_title = "alle Reservierungen";
-    require "../templates/header.php";
+    require "../_templates/header.php";
     if ($userIsAdmin) {
         require "../reservations/_reservations.php";
     }
-    require "../templates/footer.php";
+    require "../_templates/footer.php";
 ?>
