+++ login/denyNoLogin.php
@@ -1,7 +1,7 @@
 <?php
     require "startSession.php";
     if (!$loggedIn) {
-        require "../templates/header.php";
+        require "../_templates/header.php";
         require "../error/not-logged-in.php";
     }
 ?>
