+++ reservations/index.php
@@ -4,9 +4,9 @@
     $page_title = "alle Reservierungen";
     require "../login/denyNoLogin.php";
     if ($loggedIn) {
-        require "../templates/header.php";
+        require "../_templates/header.php";
         $sql_untilDateCheck = "AND ((r.rdatumbis) + 1 > (CURRENT_DATE))"; //FIXME I assume that sql call works
         require "_reservations.php";
     }
-    require "../templates/footer.php";
+    require "../_templates/footer.php";
 ?>
